requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
